Add skeleton of query service
