+++ crates/ghss_store/src/db/read.rs
@@ -7,8 +7,9 @@ pub struct DB {
 }
 
 impl DB {
-    pub fn open(name: String) -> Result<DB> {
-        let conn = Connection::open_with_flags(name, OpenFlags::SQLITE_OPEN_READ_ONLY)?;
+    pub fn open(directory: &str, repository_id: &str) -> Result<DB> {
+        let path = format!("{}/{}.db", directory, repository_id);
+        let conn = Connection::open_with_flags(path, OpenFlags::SQLITE_OPEN_READ_ONLY)?;
         Ok(DB { conn })
     }
